+++ src/shapes/metrics_path.cpp
@@ -189,7 +189,7 @@ float MetricsPath::computeLength(const Mat2D& transform)
 				const Vec2D& toIn = pen[2];
 				const Vec2D& to = pen[3];
 
-				int index = m_CubicSegments.size();
+				int index = (int) m_CubicSegments.size();
 				float partLength = segmentCubic(
 				    from, fromOut, toIn, to, 0.0f, 0.0f, 1.0f, m_CubicSegments);
 				m_Lengths.push_back(partLength);
@@ -218,7 +218,7 @@ void MetricsPath::trim(float startLength,
 	// We need to find the first part to trim.
 	float length = 0.0f;
 
-	int partCount = m_Parts.size();
+	int partCount = (int) m_Parts.size();
 	int firstPartIndex = -1, lastPartIndex = partCount - 1;
 	float startT = 0.0f, endT = 1.0f;
