+++ src/shapes/paint/trim_path.cpp
@@ -45,7 +45,7 @@ RenderPath* TrimPath::effectPath(MetricsPath* source)
 				endLength -= totalLength;
 			}
 
-			int i = 0, subPathCount = subPaths.size();
+			int i = 0, subPathCount = (int) subPaths.size();
 			while (endLength > 0)
 			{
 				MetricsPath* path =
